Applied changes for update to BOSS v3 API.
I haven't yet built a 64 bit DLL, so that's not committed yet.

There is an anomalous call to Get/SetLoadOrder in bosh.py, line 10763
(def rename), which looks like it probably crept in when Valda last
rebased off Wrye Bash, as there's not a single other use of BAPIv2's
load order features in the code.

That call will  have to be replaced, or renaming won't work, so I need
to have a look at what the usual code is for changing load order.